+++ kycho/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: kycho <kycho@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/25 13:46:49 by kycho             #+#    #+#             */
-/*   Updated: 2021/07/06 11:28:02 by kycho            ###   ########.fr       */
+/*   Updated: 2021/07/06 15:12:44 by kycho            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ public:
 	std::string					root;
 	std::vector<std::string>	index;
 	bool						autoindex;
-	unsigned long long 			client_max_body_size;
+	unsigned long 				client_max_body_size;
 	std::map<int, std::string>	error_page;
 	// return_value;
 	
@@ -52,10 +52,6 @@ public:
 	void print_status_for_debug(std::string prefix);  // TODO : remove
 
 private:
-	// struct compare_uri_for_descending_order_by_length
-	// {
-	// 	bool operator()(const Location* first, const Location* second) const; 
-	// }; 
 	
 	static bool compare_uri_for_descending_order_by_length(const Location* first, const Location* second);
 };
